backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
websockets==10.4
pydantic==1.10.13
email-validator==1.3.1
opencv-python==4.8.0.74
numpy==1.26.4
pillow==10.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.0
